Fixed player bleed when player logs
